% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{crps_rf}
\alias{crps_rf}
\title{This function calculates the CRPS for an Input of the form (observations, points, predicted cdf).}
\usage{
crps_rf(y, points, cdf)
}
\arguments{
\item{y}{a vector of observations, or a scalar (in this case, the y value will be used for all predictions)}

\item{points}{where the cdfs jump}

\item{cdf}{}
}
\value{
A vector of CRPS values
}
\description{
Computes the CRPS of raw forecasts.
}
\details{
This function uses adapted code taken from the function \code{cprs.idr} of the \pkg{isodistrreg} package.
}
