% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{e_value_calculate_lambda_for_alternative_betting_each}
\alias{e_value_calculate_lambda_for_alternative_betting_each}
\title{Betting strategies}
\usage{
e_value_calculate_lambda_for_alternative_betting_each(
  T.F.G,
  crps.F.para,
  crps.G.para,
  inf.crps,
  suffix,
  F.proportion,
  G.proportion,
  crps.alt.old,
  k
)
}
\arguments{
\item{T.F.G}{= (crps.F - crps.G) / inf.crps}

\item{crps.F.para}{of the form \code{list("mu" = mu, "sd" = 1)} or \code{list("points.cdf" = tibble::tibble with points and cdfs)}}

\item{crps.G.para}{of the form \code{ist("mu" = -mu, "sd" = 1)} or \code{list("points.cdf" = tibble::tibble with points and cdfs)}}

\item{inf.crps}{= infimum of crps.F - crps.G over y}

\item{suffix}{defines the name of the betting strategy, it returns a list then, with the respectively names.suffix}

\item{F.proportion}{F.proportion * crps.F.para$sample.fun + G.proportion * crps.G.para$sample.fun}

\item{G.proportion}{F.proportion * crps.F.para$sample.fun + G.proportion * crps.G.para$sample.fun}

\item{crps.alt.old}{is the return value of the last call to \code{\link{e_value}} for the specific betting strategy}

\item{k}{is the parameter to specify how many new observations are evaulated, used for  sequential run of \code{e_value}}
}
\value{
Returns a list for each alternative betting strategy specified, containing \code{e.value.alt.(suffix)},
\code{e.value.alt.(suffix).prod} and \code{lambda.alt.(suffix)}.
\code{e.value.alt.(suffix)} and \code{lambda.alt.(suffix)} are both vectors of the size of
the observation. \code{e.value.alt.(suffix).prod} is the e-value process at time k (number of observations).
}
\description{
This method calculates the lambda with the alternative betting strategy for each strategy, or can be used to calculate
individual proportions of the combinations of F and G.
}
\seealso{
\code{\link{e_value}}
}
