% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{e_value_calculate_lambda_for_grapa_betting}
\alias{e_value_calculate_lambda_for_grapa_betting}
\title{This method calculates the lambda with the grapa betting strategy}
\usage{
e_value_calculate_lambda_for_grapa_betting(T.F.G)
}
\arguments{
\item{T.F.G}{= (crps.F - crps.G) / inf.crps}
}
\description{
This method calculates the lambda with the grapa betting strategy
}
