% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{filter_tibble_for_usual_forecasts_and_prettify}
\alias{filter_tibble_for_usual_forecasts_and_prettify}
\title{This method filters the outcome of the [sim_e_values()] for the usual forecasts and relocates the p-value-column to the end.}
\usage{
filter_tibble_for_usual_forecasts_and_prettify(dt)
}
\arguments{
\item{dt}{is the result of the [sim_e_values(usual.forecasts = TRUE)]. But it has to have the usual.forecasts = TRUE.}
}
\description{
This method filters the outcome of the [sim_e_values()] for the usual forecasts and relocates the p-value-column to the end.
}
