% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{getFile}
\alias{getFile}
\title{This method returns the R object save it the file with getwd() + filename.}
\usage{
getFile(filename)
}
\arguments{
\item{filename}{is the name of the file.}
}
\value{
the R object in the file.
}
\description{
This method returns the R object save it the file with getwd() + filename.
}
\examples{
\dontrun{getFile("/test.rds")}
}
