% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{print_crps_diff_histogram_usual_forecasts}
\alias{print_crps_diff_histogram_usual_forecasts}
\title{This method creates a pdf with the plots of the histograms of the difference of the crps values.}
\usage{
print_crps_diff_histogram_usual_forecasts(dt, filename)
}
\arguments{
\item{dt}{is the result of the [sim_e_values(usual.forecasts = TRUE)]. But it has to have the usual.forecasts = TRUE.}

\item{filename}{to save the pdf to.}
}
\description{
This method creates a pdf with the plots of the histograms of the difference of the crps values.
}
\examples{
print_crps_diff_histogram_usual_forecasts(dt, "/test")
}
