% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{print_e_values_histogram_usual_forecasts}
\alias{print_e_values_histogram_usual_forecasts}
\title{This method creates a pdf with the plots of the histograms of the overall time difference of the e-values.}
\usage{
print_e_values_histogram_usual_forecasts(dt)
}
\arguments{
\item{dt}{is the result of the [sim_e_values(usual.forecasts = TRUE)]. But it has to have the usual.forecasts = TRUE.}
}
\description{
This method creates a pdf with the plots of the histograms of the overall time difference of the e-values.
}
