% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{print_rej_rate_perfect_loosing_power}
\alias{print_rej_rate_perfect_loosing_power}
\title{This method creates a pdf with the plots of the rejection rate of the forecasts with bias in the mean or variance.}
\usage{
print_rej_rate_perfect_loosing_power(dt)
}
\arguments{
\item{dt}{is the result of the [sim_e_values(loosing.power.forecasts = TRUE)]. But it has to have the loosing.power.forecasts = TRUE.}
}
\description{
This method creates a pdf with the plots of the rejection rate of the forecasts with bias in the mean or variance.
}
