% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{printPlot}
\alias{printPlot}
\title{This method prints the plots of a method and saves them to a file.}
\usage{
printPlot(fileName, fun, data, paper = "A4")
}
\arguments{
\item{fileName}{of the resulting file.}

\item{fun}{= \(x) {function to be called (x) }, can be any function that produces an output.}

\item{data}{is the data to be put into fun. Usually return of [sim_e_values()].}

\item{paper}{= "A4", can be replaced with "A4r" for rotated A4.}
}
\description{
This method prints the plots of a method and saves them to a file.
}
\examples{
\dontrun{printPlot("test", \(x) {print_crps_diff_histogram_usual_forecasts(x)}, dt)}
}
