% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{rcdf_rf}
\alias{rcdf_rf}
\title{Random values of the cumulative distribution function (CDF) of raw forecasts}
\usage{
rcdf_rf(points.cdf, n)
}
\arguments{
\item{points.cdf}{a \code{data.frame} of numeric variables, used to compute the empirical distribution of the variables in \code{points.cdf}.}

\item{n}{is the number of randomly generated thresholds to evaluate the cdf at.}
}
\value{
A vector of probabilities giving the evaluated CDFs at the randomly generated thresholds.
}
\description{
Evaluate the cumulative distribution function (CDF) of raw forecasts in a \code{data.frame} at randomly generated thresholds.
}
\details{
The CDFs are considered as piecewise constant stepfunctions. The \code{points} in the \code{data.frame} \code{points.cdf} are the points where the empirical distribution of the forecasts has jumps and \code{cdf} in the \code{data.frame} \code{points.cdf} are the corresponding CDF values. This function uses adapted code taken from the function \code{cdf} of the \pkg{isoditrreg} package.
}
\seealso{
\code{\link{e_value}}
}
