% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{sim_e_values}
\alias{sim_e_values}
\title{This method performs an simulation for e-values and their power.}
\usage{
sim_e_values(
  n.it = 200,
  n.obs = 100,
  loosing.power.only = FALSE,
  method = list("GRAPA", "lambda", "alternative"),
  lambda = 0.5,
  p.value.method = "t",
  forecasts.input = NA,
  loosing.power.forecasts = loosing.power.only,
  usual.forecasts = TRUE,
  file.folder = getwd()
)
}
\arguments{
\item{n.it}{= 200, number of iterations the simulation should make.}

\item{n.obs}{= 100, number of simulated observation in each iteration step.}

\item{loosing.power.only}{= FALSE, if "TRUE" only the e-values for comparing forecasts with a bias in the mean or variance are calculated.}

\item{method}{= list("GRAPA", "lambda", "alternative", "alternative-mean"), is a list containing all the method names for calculating the different lambdas.}

\item{lambda}{= 0.5, lambda entry for a fixed value.}

\item{p.value.method}{= "t", can be "t" for t-test and "dm" for dm.test of the package forecast.}

\item{forecasts.input}{= NA, one can pass a list for different forecasts to be taken into account for calculating the e-values, they have to be of the form list("additional-forecast" = list("mu" = mu, "sd" = 1)) or for mixed forecasts use "unfocused" = list("mu" = cbind(mu, mu + tau), "sd" = matrix(nrow = n.obs, ncol = 2, 1), "w" = matrix(nrow = n.obs, ncol = 2, 1 / 2)). Use [forecast_input()] to produce correct form.}

\item{loosing.power.forecasts}{= loosing.power.only, set to TRUE if you want to include the predefined forecasts with a bias in the mean or variance.}

\item{usual.forecasts}{= TRUE, set to FALSE if you want to exclude the predefined usual forecasts.}

\item{file.folder}{= getwd(), where to save the output to.}
}
\description{
This method performs an simulation for e-values and their power.
}
