% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosing_power_sim.R
\name{special_forecasts}
\alias{special_forecasts}
\title{This method provides predefined forecasts.}
\usage{
special_forecasts(
  mu,
  tau,
  n.obs,
  loosing.power.forecasts = FALSE,
  usual.forecasts = TRUE
)
}
\description{
This method provides predefined forecasts.
}
