% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loosing_power_sim.R
\name{special_forecasts_logs}
\alias{special_forecasts_logs}
\title{This method provides predefined forecasts for the log score.}
\usage{
special_forecasts_logs(
  mu,
  tau,
  n.obs,
  loosing.power.forecasts = FALSE,
  usual.forecasts = TRUE
)
}
\description{
This method provides predefined forecasts for the log score.
}
