% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{crps_rf}
\alias{crps_rf}
\title{Continuous ranked probability score (CRPS)}
\usage{
crps_rf(y, points.cdf)
}
\arguments{
\item{y}{a numeric vector of observations of the same length as the number of points.cdf, or of length 1.}

\item{points.cdf}{a \code{data.frame} of numeric variables, used to compute the empirical distribution of the
variables in \code{points.cdf}.}
}
\value{
If the input y is only a scalar, then it does return a single CRPS value. If y is a vector, it does return a vector
of CRPS values, evaluated for each y.
}
\description{
Computs the CRPS of raw forecasts.
}
\details{
This function uses adapted code taken from the function \code{crps_edf} of the \pkg{scoringRules} package and of the
function \code{cdf} of the \pkg{isoditrreg} package.
}
\seealso{
\code{\link{e_value}}
}
