% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{e_value}
\alias{e_value}
\title{This method is the main method of this package and calculates the e.values for one iterationstep.
This method uses the optional stopping principle, because of max(cumprod(e.value).}
\usage{
e_value(
  y,
  crps.F.para,
  crps.G.para,
  idx = 1,
  method = list("alt-cons"),
  lambda = 0.5,
  p.value.method = NA
)
}
\arguments{
\item{y}{oberservations y \in R^k}

\item{crps.F.para}{of the form list("mu" = mu, "sd" = 1), here mu and sd do not have to be of the same form, but for mixnorm, they all have to have the same dimensions}

\item{crps.G.para}{of the form list("mu" = mu, "sd" = 1), here mu and sd do not have to be of the same form, but for mixnorm, they all have to have the same dimensions}

\item{idx}{= 1, this is an aditional parameter to idex the run of this call}

\item{method}{= list("GRAPA", "lambda", "alt-conf", "alt-cons", "alt-more-cons"), is a list containing all the method names for calculating the different lambdas}

\item{lambda}{= 0.5, lambda entry for a fixed value}

\item{p.value.method}{= NA, can be "t" for t-test and "dm" for dm.test of the package forecast}

\item{mu}{is the mean parameter for y ~ N(mu, 1)}
}
\description{
This method is the main method of this package and calculates the e.values for one iterationstep.
This method uses the optional stopping principle, because of max(cumprod(e.value).
}
