% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{e_value_calculate_lambda_for_alternative_betting}
\alias{e_value_calculate_lambda_for_alternative_betting}
\title{This method calculates the lambda with the alternative betting strategy}
\usage{
e_value_calculate_lambda_for_alternative_betting(
  T.F.G,
  crps.F.para,
  crps.G.para,
  inf.crps,
  method
)
}
\arguments{
\item{T.F.G}{= (crps.F - crps.G) / inf.crps}

\item{crps.F.para}{of the form list("mu" = mu, "sd" = 1)}

\item{crps.G.para}{of the form list("mu" = -mu, "sd" = 1)}

\item{inf.crps}{= infimum of crps.F - crps.G over y}

\item{method}{= list("GRAPA", "lambda", "alternative", "alternative-mean"), is a list containing all the method names for calculating the different lambdas, only parameter used in this method is 'alternative-mean'}
}
\description{
This method calculates the lambda with the alternative betting strategy
}
