% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{forecast_input}
\alias{forecast_input}
\title{Creates list for one input forecast of the form N(mu,sd)}
\usage{
forecast_input(mu, sd, w = NA)
}
\arguments{
\item{mu}{is the mean of the forecast, can be constant, vector or matrix.}

\item{sd}{is the variance of the forecast, can be a constant, vector or matrix.}

\item{w}{= NA, optional, is the weight of a mixed forecast, can be a constant, vector or matrix.}

\item{forecast.name}{is the name of the forecast.}
}
\description{
This is a helper function which returns the correct form for the forecast input for [sim_e_values()].
}
