% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{get_inf_crps}
\alias{get_inf_crps}
\title{This method calculates the infimum of crps.F - crps.G.}
\usage{
get_inf_crps(crps.F.para, crps.G.para, n.obs, k = NA, old.inf = NA)
}
\arguments{
\item{crps.F.para}{of the form list("mu" = mu, "sd" = 1)}

\item{crps.G.para}{of the form list("mu" = -mu, "sd" = 1)}
}
\description{
This method calculates the infimum of crps.F - crps.G.
}
