% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_scripts.R
\name{optim_inf_value}
\alias{optim_inf_value}
\title{Calculating Infimum of difference of two forecasts}
\usage{
optim_inf_value(f, start.points = 2, min.value = 1e-04, max.value = 1)
}
\arguments{
\item{f}{the function which has to be optimized. The function has to take one value, ie f <- \(y) {}.}

\item{start.points}{= 2, how many points to start the grid search with}

\item{min.value}{= 0.0001, minimal value for the start.points to choose from}

\item{max.value}{= 1, maximum value for the start.points to choose from}
}
\value{
Returns a single infimum value.
}
\description{
This method makes a gitter search for global minima. The start points are taken in between the
\code{min.value} and the \code{max.value}.
}
\seealso{
\code{\link{e_value}}
}
